requirements.txt
numpy~=2.2.4
pandas~=2.2.3
pandas-ta==0.3.14b0
matplotlib~=3.10.1
scikit-learn~=1.6.1
tensorflow==2.14.0
MetaTrader5~=5.0.4874
joblib==1.3.2
requests~=2.32.3
scikit-optimize==0.9.0
PyYAML==6.0.1
tqdm~=4.66.3
torch~=2.6.0
pandas_ta~=0.3.14b0
yfinance~=0.2.55
PyInquirer==1.0.3
questionary~=2.1.0
seaborn~=0.13.2
xgboost~=3.0.0
scikit-optimize~=0.9.0
